add(kustomize): config tree - separate from deployment for clarity
we do not merge the deployent and config kustomize trees, so that they
remain clearly distinguishable at the ./deploy level.

This implicitly, yet clearly, indicates that users might want to
kustomize above all the config train. And following the path to the
source manifests at ./service/.../k8s, they have a clear understanding
how each service is supposed to be configured and work.