ref(kustomize): move service manifests in subfolder
using a subfolder we optain the feature to chain maniests and their
config together in parallel. Passing this distinction through to the
top level bundel allows for a unified kustomize overarride of the
many config (Secret) resources, instead of merging them artifially
together in one. THis also reduces the overlay delta for dev
environments since the secret names do not change. see next commit